sunaudio: Implement replacement for `sunau` module
The `sunau` module is being removed in Python 3.13, however, we use this
format in `slowrxd` because it's dead simple to implement, especially if
you just want a single format.

While we could implement `.wav` format (Microsoft RIFF), it is a
slightly more complicated format, which would need to be added to
`slowrxd`.  Easier to just apply the KISS principle.